ITIL Timeline feature
Add followups to changes and problems; closes #3222
Add followup search options to Changes and Problems
Add approbation
Add timeline to Changes and Problems fixes #3223
Drop tabs replaced by timeline
Update schema for new timeline items
Merge ticket, change, and problem followups
Factorize getItilActors
Add tests
