+++ CHANGELOG.md
@@ -6,7 +6,13 @@ The present file will list all changes made to the project; according to the
 ## [9.4] unreleased
 
 ### Added
+
 - Ability to link project with problems and tickets.
+- Add followups to Changes and Problems
+- Add timeline to Changes and Problems
+- Fix styles, search, and tests
+- Ability to link project with problems and tickets
+- TicketFollowup changed to ITILFollowup
 - Search on devices from Printers and Network equipments
 
 ### Changed
@@ -22,6 +28,10 @@ The present file will list all changes made to the project; according to the
 - Remove `$CFG_GLPI['use_rich_text']` parameter. Will now be `true` per default.
 - Remove `$CFG_GLPI['ticket_timeline']` parameter. Will now be `true` per default.
 - Remove `$CFG_GLPI['ticket_timeline_keep_replaced_tabs']` parameter. Will now be `false` per default.
+- Move `TicketFollowup::showFormMassiveAction` to `ITILFollowup` and deprecate.
+- Move `TicketFollowup::showMassiveActionsSubForm` to `ITILFollowup` and deprecate.
+- Move `TicketFollowup::processMassiveActionsForOneItemtype` to `ITILFollowup` and deprecate.
+- Move `TicketFollowup::showFormMassiveAction` to `ITILFollowup` and deprecate.
 
 #### Removed
