+++ front/itilfollowup.form.php
@@ -36,24 +36,29 @@ include ('../inc/includes.php');
 
 Session::checkLoginUser();
 
-$fup = new TicketFollowup();
+$fup = new ITILFollowup();
+
+if (!isset($_POST['itemtype']) || !class_exists($_POST['itemtype'])) {
+   Html::displayErrorAndDie('Lost');
+}
+$track = new $_POST['itemtype'];
+
 
 if (isset($_POST["add"])) {
    $fup->check(-1, CREATE, $_POST);
    $fup->add($_POST);
 
-   Event::log($fup->getField('tickets_id'), "ticket", 4, "tracking",
+   Event::log($fup->getField('items_id'), strtolower($_POST['itemtype']), 4, "tracking",
               //TRANS: %s is the user login
               sprintf(__('%s adds a followup'), $_SESSION["glpiname"]));
    Html::back();
 
 } else if (isset($_POST['add_close'])
            ||isset($_POST['add_reopen'])) {
-   $ticket = new Ticket();
-   if ($ticket->getFromDB($_POST["tickets_id"]) && $ticket->canApprove()) {
+   if ($track->getFromDB($_POST['items_id']) && (method_exists($track, 'canApprove') && $track->canApprove())) {
       $fup->add($_POST);
 
-      Event::log($fup->getField('tickets_id'), "ticket", 4, "tracking",
+      Event::log($fup->getField('items_id'), strtolower($_POST['itemtype']), 4, "tracking",
                  //TRANS: %s is the user login
                  sprintf(__('%s approves or refuses a solution'), $_SESSION["glpiname"]));
       Html::back();
@@ -63,19 +68,19 @@ if (isset($_POST["add"])) {
    $fup->check($_POST['id'], UPDATE);
    $fup->update($_POST);
 
-   Event::log($fup->getField('tickets_id'), "ticket", 4, "tracking",
+   Event::log($fup->getField('items_id'), strtolower($_POST['itemtype']), 4, "tracking",
               //TRANS: %s is the user login
               sprintf(__('%s updates a followup'), $_SESSION["glpiname"]));
-   Html::redirect(Toolbox::getItemTypeFormURL('Ticket')."?id=".$fup->getField('tickets_id'));
+   Html::redirect($track->getFormURL()."?id=".$fup->getField('items_id'));
 
 } else if (isset($_POST["purge"])) {
    $fup->check($_POST['id'], PURGE);
    $fup->delete($_POST, 1);
 
-   Event::log($fup->getField('tickets_id'), "ticket", 4, "tracking",
+   Event::log($fup->getField('items_id'), strtolower($_POST['itemtype']), 4, "tracking",
               //TRANS: %s is the user login
               sprintf(__('%s purges a followup'), $_SESSION["glpiname"]));
-   Html::redirect(Toolbox::getItemTypeFormURL('Ticket')."?id=".$fup->getField('tickets_id'));
+   Html::redirect($track->getFormURL()."?id=".$fup->getField('items_id'));
 }
 
 Html::displayErrorAndDie('Lost');
