+++ front/problem.form.php
@@ -107,6 +107,19 @@ if (isset($_POST["add"])) {
               //TRANS: %s is the user login
               sprintf(__('%s adds an actor'), $_SESSION["glpiname"]));
    Html::redirect($problem->getFormURLWithID($_POST['problems_id']));
+} else if (isset($_REQUEST['delete_document'])) {
+   $doc = new Document();
+   $doc->getFromDB(intval($_REQUEST['documents_id']));
+   if ($doc->can($doc->getID(), UPDATE)) {
+      $document_item = new Document_Item;
+      $found_document_items = $document_item->find("itemtype = 'Problem' ".
+                                                   " AND items_id = ".intval($_REQUEST['problems_id']).
+                                                   " AND documents_id = ".$doc->getID());
+      foreach ($found_document_items  as $item) {
+         $document_item->delete(Toolbox::addslashes_deep($item), true);
+      }
+   }
+   Html::back();
 } else {
    Html::header(Problem::getTypeName(Session::getPluralNumber()), $_SERVER['PHP_SELF'], "helpdesk", "problem");
    $problem->display($_GET);
