+++ inc/itilsolution.class.php
@@ -294,18 +294,16 @@ class ITILSolution extends CommonDBChild {
 
       //default status for global solutions
       $status = CommonITILValidation::ACCEPTED;
-      if ($input['itemtype'] == Ticket::getType()) {
-         $autoclosedelay =  Entity::getUsedConfig(
-            'autoclose_delay',
-            $this->item->getEntityID(),
-            '',
-            Entity::CONFIG_NEVER
-         );
+      $autoclosedelay =  Entity::getUsedConfig(
+         'autoclose_delay',
+         $this->item->getEntityID(),
+         '',
+         Entity::CONFIG_NEVER
+      );
 
-         // 0 = immediatly
-         if ($autoclosedelay != 0) {
-            $status = CommonITILValidation::WAITING;
-         }
+      // 0 = immediatly
+      if ($autoclosedelay != 0) {
+         $status = CommonITILValidation::WAITING;
       }
 
       //Auto approval; store user and date
@@ -333,19 +331,18 @@ class ITILSolution extends CommonDBChild {
 
       $item = $this->item;
       $status = $item::CLOSED;
-      if ($this->item->getType() == 'Ticket') {
-         $autoclosedelay =  Entity::getUsedConfig(
-            'autoclose_delay',
-            $this->item->getEntityID(),
-            '',
-            Entity::CONFIG_NEVER
-         );
+      $autoclosedelay =  Entity::getUsedConfig(
+         'autoclose_delay',
+         $this->item->getEntityID(),
+         '',
+         Entity::CONFIG_NEVER
+      );
 
          // 0 = immediatly
-         if ($autoclosedelay != 0) {
-            $status = $item::SOLVED;
-         }
+      if ($autoclosedelay != 0) {
+         $status = $item::SOLVED;
       }
+
       $this->item->update([
          'id'     => $this->item->getID(),
          'status' => $status
@@ -426,9 +423,9 @@ class ITILSolution extends CommonDBChild {
       }
 
       // show approbation form on top when ticket is solved
-      if ($item instanceof Ticket && $item->canApprove()) {
+      if ($item->canApprove()) {
          echo "<div class='approbation_form'>";
-         $followup_obj = new TicketFollowup();
+         $followup_obj = new ITILFollowup();
          $followup_obj->showApprobationForm($item);
          echo "</div>";
       }
