+++ inc/problem.class.php
@@ -158,6 +158,19 @@ class Problem extends CommonITILObject {
    }
 
 
+   /**
+    * is the current user could reopen the current problem
+    * @since  9.2
+    * @return boolean
+    */
+   function canReopen() {
+      return Session::haveRight('followup', CREATE)
+             && in_array($this->fields["status"], $this->getClosedStatusArray())
+             && ($this->isAllowedStatus($this->fields['status'], self::INCOMING)
+                 || $this->isAllowedStatus($this->fields['status'], self::ASSIGNED));
+   }
+
+
    function pre_deleteItem() {
       global $CFG_GLPI;
 
@@ -174,7 +187,13 @@ class Problem extends CommonITILObject {
          $nb = 0;
          switch ($item->getType()) {
             case __CLASS__ :
-               $ong = [1 => __('Analysis')];
+               $timeline    = $item->getTimelineItems();
+               $nb_elements = count($timeline);
+
+               $ong = [
+                  5 => __("Processing problem")." <sup class='tab_nb'>$nb_elements</sup>",
+                  1 => __('Analysis')
+               ];
 
                if ($item->canUpdate()) {
                   $ong[4] = __('Statistics');
@@ -206,6 +225,13 @@ class Problem extends CommonITILObject {
                case 4 :
                   $item->showStats();
                   break;
+               case 5 :
+                  echo "<div class='timeline_box'>";
+                  $rand = mt_rand();
+                  $item->showTimelineForm($rand);
+                  $item->showTimeline($rand);
+                  echo "</div>";
+                  break;
             }
       }
       return true;
@@ -217,17 +243,14 @@ class Problem extends CommonITILObject {
       // show related tickets and changes
       $ong = [];
       $this->addDefaultFormTab($ong);
-      $this->addStandardTab('ITILSolution', $ong, $options);
+      $this->addStandardTab(__CLASS__, $ong, $options);
       $this->addStandardTab('Problem_Ticket', $ong, $options);
       $this->addStandardTab('Change_Problem', $ong, $options);
-      $this->addStandardTab('ProblemTask', $ong, $options);
       $this->addStandardTab('ProblemCost', $ong, $options);
       $this->addStandardTab('Itil_Project', $ong, $options);
       $this->addStandardTab('Item_Problem', $ong, $options);
       $this->addStandardTab('Change_Problem', $ong, $options);
       $this->addStandardTab('Problem_Ticket', $ong, $options);
-      $this->addStandardTab(__CLASS__, $ong, $options);
-      $this->addStandardTab('Document_Item', $ong, $options);
       $this->addStandardTab('Notepad', $ong, $options);
       $this->addStandardTab('KnowbaseItem_Item', $ong, $options);
       $this->addStandardTab('Log', $ong, $options);
@@ -242,6 +265,13 @@ class Problem extends CommonITILObject {
       $DB->delete(
          'glpi_problemtasks', [
             'problems_id'  => $this->fields['id']
+         ]
+      );
+
+      $DB->delete(
+         'glpi_itilfollowups', [
+            'itemtype'  => 'Problem',
+            'items_id'  => $this->fields['id']
          ]
       );
 
@@ -488,6 +518,8 @@ class Problem extends CommonITILObject {
       ];
 
       $tab = array_merge($tab, Notepad::rawSearchOptionsToAdd());
+
+      $tab = array_merge($tab, ITILFollowup::rawSearchOptionsToAdd());
 
       $tab = array_merge($tab, ProblemTask::rawSearchOptionsToAdd());
 
@@ -1283,6 +1315,7 @@ class Problem extends CommonITILObject {
    }
 
 
+
    static function getCommonSelect() {
 
       $SELECT = "";
@@ -1511,12 +1544,12 @@ class Problem extends CommonITILObject {
    /**
     * Number of tasks of the problem
     *
-    * @return followup count
+    * @return task count
    **/
    function numberOfTasks() {
       global $DB;
 
-      // Set number of followups
+      // Set number of tasks
       $query = "SELECT COUNT(*)
                 FROM `glpi_problemtasks`
                 WHERE `problems_id` = '".$this->fields["id"]."'";
@@ -1541,5 +1574,4 @@ class Problem extends CommonITILObject {
 
       return $values;
    }
-
 }
