+++ inc/toolbox.class.php
@@ -1884,6 +1884,7 @@ class Toolbox {
                            // Tasks or Followups
                            $forcetab = str_replace( 'TicketFollowup$1', 'Ticket$1', $forcetab);
                            $forcetab = str_replace( 'TicketTask$1', 'Ticket$1', $forcetab);
+                           $forcetab = str_replace( 'ITILFollowup$1', 'Ticket$1', $forcetab);
                            Html::redirect(Ticket::getFormURLWithID($data[1])."&$forcetab");
 
                         } else if (!empty($data[0])) { // redirect to list
@@ -1941,6 +1942,7 @@ class Toolbox {
                               // force redirect to timeline when timeline is enabled
                               $forcetab = str_replace( 'TicketFollowup$1', 'Ticket$1', $forcetab);
                               $forcetab = str_replace( 'TicketTask$1', 'Ticket$1', $forcetab);
+                              $forcetab = str_replace( 'ITILFollowup$1', 'Ticket$1', $forcetab);
                               Html::redirect($item->getFormURLWithID($data[1])."&$forcetab");
                            }
